Found axis bug in the simulator

+++ Simulator/2017MockWpi/snobot_sim/com/snobot/simulator/gui/joysticks/sub_panels/AnalogControllerInputPanel.java
@@ -19,8 +19,8 @@ public class AnalogControllerInputPanel extends JPanel
         add(mSlider);
     }
 
-    public void setValue(int aValue)
+    public void setValue(double aValue)
     {
-        mSlider.setValue(aValue);
+        mSlider.setValue((int) aValue);
     }
 }
